Moves the package init function to its own files
Conditionally built by the presence of the gowslmock build tag
