+++ windows-agent/internal/daemon/init_windows.go
@@ -1,0 +1,10 @@
+//go:build !gowslmock
+
+package daemon
+
+func init() {
+	defaultOptions = options{
+		wslSystemCmd:         []string{"wsl.exe", "--system", "wslinfo", "--networking-mode", "-n"},
+		getAdaptersAddresses: getWindowsAdaptersAddresses,
+	}
+}
